Added File checking to Task 7.1

+++ Task7/Program.cs
@@ -4,17 +4,24 @@ string output = $"{Environment.GetFolderPath(System.Environment.SpecialFolder.Us
 string input = $"{Environment.GetFolderPath(System.Environment.SpecialFolder.UserProfile)}/data.txt";
 
 Console.WriteLine($"Writing data from {input} to {output}...");
-Console.Write("Using StreamReader/Writter... ");
-using (StreamReader reader = new(input))
+if (File.Exists(input))
 {
-    using StreamWriter writer = new(output);
-    writer.Write(reader.ReadToEnd());
-}
-Console.WriteLine("Done.");
+    Console.Write("Using StreamReader/Writter... ");
+    using (StreamReader reader = new(input))
+    {
+        using StreamWriter writer = new(output);
+        writer.Write(reader.ReadToEnd());
+    }
+    Console.WriteLine("Done.");
 
-Console.Write("Using File.WriteAllText... ");
-using (StreamReader reader = new(input))
-{
-    File.WriteAllText(output, reader.ReadToEnd());
+    Console.Write("Using File.WriteAllText... ");
+    using (StreamReader reader = new(input))
+    {
+        File.WriteAllText(output, reader.ReadToEnd());
+    }
+    Console.WriteLine("Done.");
 }
-Console.WriteLine("Done.");
+else
+{
+    Console.WriteLine($"Uhhhhh... {input} doesn't exist.");
+}
